ABM estado, prioridad y persona
