+++ src/main/java/com/example/Taller/Entity/UsuarioEntity.java
@@ -15,6 +15,18 @@ public class UsuarioEntity {
     @ManyToOne(optional = false)
     private TipoUsuarioEntity tipoUsuario;
 
+    @Column(nullable = false)
+    private String nombre;
+
+    @Column(nullable = false)
+    private String apellido;
+
+    @Column(nullable = false)
+    private String telefono;
+
+    @Column(nullable = false, unique = true)
+    private String cedula;
+
     public String getEmail() {
         return email;
     }
@@ -37,5 +49,37 @@ public class UsuarioEntity {
 
     public void setTipoUsuario(TipoUsuarioEntity tipoUsuario) {
         this.tipoUsuario = tipoUsuario;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public String getTelefono() {
+        return telefono;
+    }
+
+    public void setTelefono(String telefono) {
+        this.telefono = telefono;
+    }
+
+    public String getCedula() {
+        return cedula;
+    }
+
+    public void setCedula(String cedula) {
+        this.cedula = cedula;
     }
 }
