--- src/main/java/com/example/Taller/Repository/PersonaRepository.java
@@ -1,8 +1,0 @@
-package com.example.Taller.Repository;
-
-import com.example.Taller.Entity.PersonaEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PersonaRepository extends JpaRepository<PersonaEntity, Integer> {
-
-}
