--- src/main/java/com/example/Taller/Service/PersonaService.java
@@ -1,16 +1,0 @@
-package com.example.Taller.Service;
-
-import com.example.Taller.Entity.PersonaEntity;
-
-import java.util.List;
-
-public interface PersonaService {
-
-    PersonaEntity guardarPersona(PersonaEntity persona);
-
-    PersonaEntity obtenerPersonaPorId(int id);
-
-    List<PersonaEntity> obtenerTodasLasPersonas();
-
-    String eliminarPersona(int id);
-}
