+++ src/main/resources/application.properties
@@ -3,5 +3,5 @@ spring.datasource.url=jdbc:mysql://localhost:3306/taller1
 spring.datasource.username=root
 spring.datasource.password=colonia31
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 server.port = 5000
